server/requirements.txt
anthropic
pyyaml